requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=10.0.0
qrcode>=7.4.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1 